test: ReservationFinishedPresenter 테스트 수정
